# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nifty 100 Analytics API",
    "description": "Read-only REST API over the Nifty 100 Analytics SQLite database (financial ratios, screener, peer comparison, valuation, reports). Sprint 6 / Day 40 deliverable.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
    "/api/v1/health": {
      "get": {
        "summary": "Health check",
        "description": "status, per-table row counts, uptime, version",
        "tags": [
          "health"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies": {
      "get": {
        "summary": "List companies",
        "description": "All 92 companies (id, name, sector, ROE, ROCE), optionally filtered",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "market_cap_category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}": {
      "get": {
        "summary": "Get company profile",
        "description": "Full company profile + latest-year KPIs. 404 if unknown.",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/pl": {
      "get": {
        "summary": "Get P&L history",
        "description": "P&L history, optional YYYY-MM year range filter",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/bs": {
      "get": {
        "summary": "Get balance sheet history",
        "description": "Balance sheet history, optional year range filter",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/cashflow": {
      "get": {
        "summary": "Get cash flow history",
        "description": "Cash flow history, optional year range filter",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to_year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/ratios": {
      "get": {
        "summary": "Get computed KPIs",
        "description": "All computed KPIs per year, or a single year if `year` is passed",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/tearsheet": {
      "get": {
        "summary": "Download tearsheet PDF",
        "description": "The pre-generated 2-page tearsheet PDF (application/pdf)",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/peers/compare": {
      "get": {
        "summary": "Get radar comparison data",
        "description": "8-axis metric values vs peer group average + benchmark company",
        "tags": [
          "companies"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/companies/{ticker}/documents": {
      "get": {
        "summary": "Get annual report links",
        "description": "Annual report links with is_url_valid (format check) flag",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/screener": {
      "get": {
        "summary": "Run the screener",
        "description": "Ranked company list filtered by any combination of thresholds. 400 for invalid params.",
        "tags": [
          "screener"
        ],
        "parameters": [
          {
            "name": "min_roe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_de",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "min_fcf",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_rev_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "min_pat_cagr_5yr",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "max_pe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          },
          "400": {
            "description": "Invalid parameter value"
          }
        }
      }
    },
    "/api/v1/sectors": {
      "get": {
        "summary": "List sectors",
        "description": "All broad sectors with company_count, median_roe, median_pe, median_de",
        "tags": [
          "sectors"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/sectors/{sector}/companies": {
      "get": {
        "summary": "Get companies in a sector",
        "description": "All companies in a broad_sector with latest-year KPIs. 404 if unknown.",
        "tags": [
          "sectors"
        ],
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/peers/{group_name}": {
      "get": {
        "summary": "Get a peer group",
        "description": "All companies in a peer group with percentile ranks. 404 if unknown.",
        "tags": [
          "peers"
        ],
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/market-cap/{ticker}": {
      "get": {
        "summary": "Get valuation multiples history",
        "description": "Historical P/E, P/B, EV/EBITDA, dividend yield, 2019-2024 (SIMULATED data)",
        "tags": [
          "valuation"
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/v1/portfolio/stats": {
      "get": {
        "summary": "Get portfolio percentile stats",
        "description": "P10-P90 percentile table for 10 core KPIs across all 92 companies",
        "tags": [
          "portfolio"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    }
  }
}